add main1.java

+++ codes/0492-Construct-the-Rectangle/Java/main1.java
@@ -1,6 +1,6 @@
 class Solution {
     public int[] constructRectangle(int area) {
-        int r = Math.sqrt(area);
+        int r = (int)Math.sqrt(area);
         while (area % r != 0) --r;
         return int[] { area / r, r };
     }
